(#332) Get rid of `internalIds` set in ChanReaderRequest. Use database to figure out whether a post is internal or not (for both threads in a catalog and posts in a thread)

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/site/parser/PostParser.java
@@ -31,6 +31,6 @@ public interface PostParser {
          * @param postNo the post id
          * @return {@code true} if referring to a post in the thread, {@code false} otherwise.
          */
-        boolean isInternal(int postNo);
+        boolean isInternal(long postNo);
     }
 }
